requirements.txt
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-mgmt-compute==27.2.0
azure-mgmt-core==1.3.2
azure-mgmt-network==21.0.1
azure-mgmt-resource==21.1.0
boto3==1.24.51
botocore==1.27.51
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
idna==3.3
isodate==0.6.1
jmespath==1.0.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.0
portalocker==2.5.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
